requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
aiosqlite>=0.19.0
apscheduler>=3.10.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
pystray>=0.19.0
plyer>=2.1.0
pyyaml>=6.0
pillow>=10.0.0
jinja2>=3.1.0
